+++ src/astar.py
@@ -4,10 +4,7 @@
 
 from graph import *
 from node import *
-from loader import *
-from visualizer import *
-from ant_colony import *
-from mtsp import *
+import numpy as np
 
 # Calculate heuristic value : distance from every city to destination city
 def heuristics(cities, destination):
@@ -85,14 +82,14 @@ def add_to_open(open, neighbor):
             return False
     return True
 
-# Convert from path detail into path route for visualization
-def detailToSimple(path):
+# Return the route information of the path
+def getRoute(path):
     route = []
     for i in range(len(path)):
         route.append(int(path[i][0]))
     return route
 
-# Return the cost of the path
+# Return the cost infromation of the path
 def getCost(path):
     return path[-1][1]
 
@@ -112,100 +109,8 @@ def generateAStarParameter(destination, maps, graph):
             if (i != j):
                 # Run the search algorithm
                 path = astar_search(graph, h, destination[i], destination[j])
-                print(path)
-                route = detailToSimple(path)
+                route = getRoute(path)
                 routes[i][j] = route
                 distance[i][j] = getCost(path)
 
     return distance, routes
-
-# Convert from simple route from A* Matrix into the real route from the maps
-def simple_into_real_route(trail, routes):
-    realPath = []
-    originCity = trail[0]
-    realOriginCity = routes[int(originCity)][1][0]
-    for i in range(len(trail)-1):
-        realPath = realPath + routes[int(originCity)][int(trail[i+1])]
-        realPath.pop()
-        originCity = trail[i+1]
-    realPath.append(realOriginCity)
-    return realPath
-
-# Change correspondent route into True
-def apply_route_on_streets(all_real_path, boolStreets):
-    for i in range(len(all_real_path)):
-        for j in range(len(all_real_path[i])-1):
-            label = str(all_real_path[i][j]) + '-' + str(all_real_path[i][j+1])
-            value = boolStreets.get(label)
-            if (value == None):
-                label = str(all_real_path[i][j+1]) + '-' + str(all_real_path[i][j])
-            boolStreets[label] = i
-
-# The main entry point for this module
-def main():
-    # Quick test for quick debugging
-    fileNode = "OL.cnode.txt"
-    fileEdge = "OL.cedge.txt"
-    numOfSalesman = 2
-
-    maps = LoadCoordinate(fileNode)
-    streets = LoadStreet(fileEdge)
-    boolStreets = GetBooleanStreets(streets)
-
-    # Create a graph
-    graph = Graph()
-
-    # Create graph connections (Actual distance)
-    graph.streetsToGraph(streets)
-
-    # Make graph undirected, create symmetric connections
-    graph.make_undirected()
-
-    all_real_path = []
-    solution = []
-    destination = [str(2000 + i * 10 + random.randint(0,70)) for i in range(numOfSalesman * 3)]
-    listOfSubDestination = splitCities(destination, numOfSalesman, destination[0], maps)
-    destinationInt = [int(destination[i]) for i in range(numOfSalesman * 3)]
-    
-    print(destination)
-
-    input()
-    for i in range(numOfSalesman):
-        MCost = []
-        MRoute = []
-        MCost, MRoute = generateAStarParameter(listOfSubDestination[i], maps, graph)
-        PrintMatrix(MCost, len(listOfSubDestination[i]))
-        print(listOfSubDestination[i])
-        print(MRoute)
-        input()
-
-        # Execute numOfSalesman TSP process and return numOfSalesman route
-        ant_colony = AntColony(np.array(MCost), 10, 5, 100, 0.95, alpha=1, beta=1)
-        shortest_path = ant_colony.run()
-        print ("Shortest path: {}".format(shortest_path))
-        if (shortest_path[0] != 'placeholder'):
-            routeParent = ConvertIntoRoute(shortest_path)
-            # ValidizeRoute(routeParent, destination)
-            print("Jarak tempuh untuk sales ke-"+str(i+1)+" adalah "+str(shortest_path[1])+" km")
-            print("Rute perjalanan untuk sales ke-"+str(i+1)+" adalah ", end="")
-            PrintRoute(routeParent)
-            real_path = simple_into_real_route(routeParent, MRoute)
-            print("Rute asli perjalanan untuk sales ke-"+str(i+1)+" adalah ", end="")
-            PrintRoute(real_path)
-            all_real_path.append(real_path)
-            solution.append([real_path, shortest_path[1]])
-        else:
-            print("Tidak terdapat rute yang memungkinkan untuk sales ke-"+str(i+1))
-            solution.append([[0], -999])
-        input()
-
-    apply_route_on_streets(all_real_path, boolStreets)
-    PrintAllSolution(solution, numOfSalesman)
-    input("Next.. to visualizer")
-
-    # Visualize the route for every salesman
-    VisualizeComplexGraph(destinationInt[0], streets, boolStreets, all_real_path, maps, len(maps), destinationInt)
-
-# Tell python to run main method
-if __name__ == "__main__": main()
-
