--- src/test.py
@@ -1,133 +1,0 @@
-
-
-# A* search
-def astar_search(graph, heuristics, start, end):
-    
-    # Create lists for open nodes and closed nodes
-    open = []
-    closed = []
-
-    # Create a start node and an goal node
-    start_node = Node(start, None)
-    goal_node = Node(end, None)
-
-    # Add the start node
-    open.append(start_node)
-    
-    # Loop until the open list is empty
-    while len(open) > 0:
-
-        # Sort the open list to get the node with the lowest cost first
-        open.sort()
-
-        # Get the node with the lowest cost
-        current_node = open.pop(0)
-
-        # Add the current node to the closed list
-        closed.append(current_node)
-        
-        # Check if we have reached the goal, return the path
-        if current_node == goal_node:
-            path = []
-            while current_node != start_node:
-                path.append(current_node.name + ': ' + str(current_node.g))
-                current_node = current_node.parent
-            path.append(start_node.name + ': ' + str(start_node.g))
-            # Return reversed path
-            return path[::-1]
-
-        # Get neighbours
-        neighbors = graph.get(current_node.name)
-
-        # Loop neighbors
-        for key, value in neighbors.items():
-
-            # Create a neighbor node
-            neighbor = Node(key, current_node)
-
-            # Check if the neighbor is in the closed list
-            if(neighbor in closed):
-                continue
-
-            # Calculate full path cost
-            neighbor.g = current_node.g + graph.get(current_node.name, neighbor.name)
-            neighbor.h = heuristics.get(neighbor.name)
-            neighbor.f = neighbor.g + neighbor.h
-
-            # Check if neighbor is in open list and if it has a lower f value
-            if(add_to_open(open, neighbor) == True):
-                # Everything is green, add neighbor to open list
-                open.append(neighbor)
-
-    # Return None, no path is found
-    return None
-
-# Check if a neighbor should be added to open list
-def add_to_open(open, neighbor):
-    for node in open:
-        if (neighbor == node and neighbor.f > node.f):
-            return False
-    return True
-
-# The main entry point for this module
-def main():
-
-    # Create a graph
-    graph = Graph()
-
-    # Create graph connections (Actual distance)
-    graph.connect('Frankfurt', 'Wurzburg', 111)
-    graph.connect('Frankfurt', 'Mannheim', 85)
-    graph.connect('Wurzburg', 'Nurnberg', 104)
-    graph.connect('Wurzburg', 'Stuttgart', 140)
-    graph.connect('Wurzburg', 'Ulm', 183)
-    graph.connect('Mannheim', 'Nurnberg', 230)
-    graph.connect('Mannheim', 'Karlsruhe', 67)
-    graph.connect('Karlsruhe', 'Basel', 191)
-    graph.connect('Karlsruhe', 'Stuttgart', 64)
-    graph.connect('Nurnberg', 'Ulm', 171)
-    graph.connect('Nurnberg', 'Munchen', 170)
-    graph.connect('Nurnberg', 'Passau', 220)
-    graph.connect('Stuttgart', 'Ulm', 107)
-    graph.connect('Basel', 'Bern', 91)
-    graph.connect('Basel', 'Zurich', 85)
-    graph.connect('Bern', 'Zurich', 120)
-    graph.connect('Zurich', 'Memmingen', 184)
-    graph.connect('Memmingen', 'Ulm', 55)
-    graph.connect('Memmingen', 'Munchen', 115)
-    graph.connect('Munchen', 'Ulm', 123)
-    graph.connect('Munchen', 'Passau', 189)
-    graph.connect('Munchen', 'Rosenheim', 59)
-    graph.connect('Rosenheim', 'Salzburg', 81)
-    graph.connect('Passau', 'Linz', 102)
-    graph.connect('Salzburg', 'Linz', 126)
-
-    # Make graph undirected, create symmetric connections
-    graph.make_undirected()
-
-    # Create heuristics (straight-line distance, air-travel distance)
-    heuristics = {}
-    heuristics['Basel'] = 204
-    heuristics['Bern'] = 247
-    heuristics['Frankfurt'] = 215
-    heuristics['Karlsruhe'] = 137
-    heuristics['Linz'] = 318
-    heuristics['Mannheim'] = 164
-    heuristics['Munchen'] = 120
-    heuristics['Memmingen'] = 47
-    heuristics['Nurnberg'] = 132
-    heuristics['Passau'] = 257
-    heuristics['Rosenheim'] = 168
-    heuristics['Stuttgart'] = 75
-    heuristics['Salzburg'] = 236
-    heuristics['Wurzburg'] = 153
-    heuristics['Zurich'] = 157
-    heuristics['Ulm'] = 0
-
-    # Run the search algorithm
-    path = astar_search(graph, heuristics, 'Frankfurt', 'Ulm')
-    print(path)
-    print()
-
-# Tell python to run main method
-if __name__ == "__main__": main()
