+++ src/visualizer.py
@@ -1,17 +1,10 @@
-# complexVisualizer.py
+# visualizer.py
 # Module to visualize the map and the streets in complex way
 # The visualization is using matplotlib library
 
 import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 import random
-
-# Check if the street is in route
-def CheckStreetInRoute(origin, destination, route):
-    for i in range (len(route)-1):
-        if ((route[i] == int(origin)) and (route[i+1] == int(destination))):
-            return True
-    return False
 
 # Plot a point on workspace
 def Plotting(P, name, k):
@@ -28,27 +21,15 @@ def Plotting(P, name, k):
 
     # If Point P is the visited city
     elif (k == 3):
-        plt.plot([x], [y], 'ro', color = "purple")
+        plt.plot([x], [y], 'ro', color = "red")
 
     # If Point P is the unvisited city
     else :
         plt.plot([x], [y], 'ro', color = "black")
+
     # label = "%d(%d,%d)" % (name, x, y)
     label = "%d" % (name)
     plt.annotate(label, (x,y), textcoords="offset points", xytext=(0,10), ha='center')
-
-# Connect two points between origin point and destination point
-def ConnectPoints(maps, origin, destination, cost, flag):
-    x1, x2 = float(maps[origin].getX()), float(maps[destination].getX())
-    y1, y2 = float(maps[origin].getY()), float(maps[destination].getY())
-
-    # If the edge between origin and destination is on the route
-    if (flag):
-        plt.plot([x1,x2],[y1,y2],'k-', color = "red")
-    
-    # The edge between origin and destination is not on the route
-    else:
-        plt.plot([x1,x2],[y1,y2],'k-', color = "black")
 
 # Connect two points
 def ConnectPointsRandomColor(maps, origin, destination, cost, k, colors):
@@ -62,7 +43,7 @@ def ConnectPointsRandomColor(maps, origin, destination, cost, k, colors):
         plt.plot([x1,x2],[y1,y2],'k-', color = "black")
 
 # Generate about numOfRoute random color to distinguish every salesman route
-def generateRandomColor(numOfRoute):
+def GenerateRandomColors(numOfRoute):
     colors = []
     for i in range(numOfRoute):
         r = random.random()
@@ -72,8 +53,8 @@ def generateRandomColor(numOfRoute):
     return colors
 
 # Visualize the graph from street, route, and length of the map information for Multiple TSP Problem
-def VisualizeComplexGraph(basecity, streets, boolStreets, listOfRoute, maps, length, destination):
-    colors = generateRandomColor(len(listOfRoute))
+def VisualizeComplexGraph(baseCity, streets, boolStreets, listOfRoute, maps, length, destination):
+    colors = GenerateRandomColors(len(listOfRoute))
     # Visualize every street in the map
     for street in streets:
         # streetRoute = -1
@@ -85,29 +66,26 @@ def VisualizeComplexGraph(basecity, streets, boolStreets, listOfRoute, maps, len
         value = boolStreets.get(label)
         streetRoute = boolStreets[label]
 
-        if (streetRoute != -1):
-            ConnectPointsRandomColor(maps, street[1], street[2], street[3], streetRoute, colors)
-        else:
-            ConnectPoints(maps, street[1], street[2], street[3], False)
+        ConnectPointsRandomColor(maps, street[1], street[2], street[3], streetRoute, colors)
 
     # Visualize every city in the map with black color
     for city in maps:
         Plotting(maps[city], float(city), 4)
 
-    # Visualize every city in the route with another color
+    # Visualize every city in the route with blue color
     for route in listOfRoute:
         for city in route:
             P = maps[str(city)]
             Plotting(P, city, 2)
 
-    # Visualize every city in the destination with another color
+    # Visualize every city in the destination with red color
     for city in destination:
         P = maps[str(city)]
         Plotting(P, city, 3)
 
     # Visualize origin city with green color
-    P = maps[str(basecity)]
-    Plotting(P, basecity, 1)
+    P = maps[str(baseCity)]
+    Plotting(P, baseCity, 1)
 
     # Create legend information based on the number of salesman
     personLegend = []
